* main:
  chore: update gas snapshot
  chore: pack `PortfolioPosition` struct
  chore: update gas snapshot
  chore: avoid performing checks for scenarios which are mutually exclusive
  chore: update gas snapshot
  chore: use branchless `min` when calculating fees
  chore: add notices on safety of `unchecked` blocks
  chore: apply new `min` function to `getPoolMaxLiquidity`
  feat: optimize `PortfolioLib.computeTau` with branchless `max`
  feat: optimize `AccountLib.debit` with branchless `min`

+++ contracts/interfaces/IPortfolio.sol
@@ -203,7 +203,7 @@ interface IPortfolioGetters {
         view
         returns (
             uint128 freeLiquidity,
-            uint256 lastTimestamp,
+            uint32 lastTimestamp,
             uint256 invariantGrowthLast,
             uint256 feeGrowthAssetLast,
             uint256 feeGrowthQuoteLast,
